[ADD] createStar in intervieweeAnswerScore
